pyproject.toml
[project]
name = "my-open-claw"
version = "0.1.0"
description = "Agentic Framework using CrewAI with Ollama (minimax-m2.5:cloud)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=0.80.0",
    "langchain-community>=0.3.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.3.0",
    "duckduckgo-search>=7.0.0",
    "Jinja2>=3.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
